Make chrome_plugin_host.cc and chrome_plugin_util.cc compile on Posix. Review URL: http://codereview.chromium.org/27150
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@10436 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/common/chrome_plugin_util.h
@@ -64,7 +64,7 @@ public:
   // Common implementation of a CPR_GetResponseInfo call.
   static int GetResponseInfo(
       const net::HttpResponseHeaders* response_headers,
-      CPResponseInfoType type, void* buf, uint32 buf_size);
+      CPResponseInfoType type, void* buf, size_t buf_size);
 };
 
 // Helper to allocate a string using the given CPB_Alloc function.
